requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
seaborn>=0.12.0
opencv-python>=4.7.0
numpy>=1.23.0
PyYAML>=6.0
